requirements.txt
numpy
scikit-learn
pdfminer.six
python-docx